+++ dcc-portal-ui/app/scripts/advanced/views/advanced.mutations.mutations.html
@@ -27,6 +27,12 @@
             <abbr data-tooltip="{{'Consequences caused by the mutation, annotated using SnpEff with Sequence Ontology terms. High, Low impact AA change is in RED or GREEN respectively.' | translate}}"><translate>Consequences</translate></abbr>
         </th>
         <th class="text-center" colspan="2"><translate># Donors affected</translate></th>
+        <th rowspan="2">
+            <i class="icon-chart-bar"
+                data-tooltip="Survival Analysis / Phenotype Comparison"
+                data-tooltip-placement="left">
+            </i>
+        </th>
     </tr>
     <tr>
         <th data-sortable data-type="mutations" data-field="affectedDonorCountFiltered" data-active="true"
@@ -80,6 +86,10 @@
                             </span>
                 </li>
             </ul>
+        </td>
+        <td>
+            <i class="icon-chart-bar phenotype-analysis"
+                data-ng-click="AdvancedCtrl.launchSurvivalAnalysis('mutation', mutation.id, mutation.id)"></i>
         </td>
     </tr>
     </tbody>
