+++ dcc-portal-ui/app/scripts/analysis/js/analysisLauncher.js
@@ -251,6 +251,8 @@
         name: gettextCatalog.getString('Pancreatic - KRAS not mutated ')
       };
 
+      console.log(p1, p2);
+
       var demoSetIds = [];
       Page.startWork();
       SetService.addSet(type, p1).then(function (r1) {
