+++ dcc-portal-ui/app/styles/styles.scss
@@ -432,7 +432,7 @@ a.no-style {
   }
 }
 
-.iobio-popup {
+.iobio-popup, .phenotype-analysis {
   color:#e9931c;
   transition: color .5s, border .5s;
   width:18px;
